Incorporate TH spec into the spec suite.
It's completely unnecessary to have two hspec suites where one can do.

With this change also namespace is slightly more polluted with
`optionsParser` function. It's not *strictly* necessary, but tests want
something easy to test and it's also can be useful for clients. (If they
want to extend generated parser or whatever.)